food is generated randomly and snake eats the food
What changes you have made?
game class is created to combine snake class and food class
game class is creating food randomly

Why this change is necessary?
to make food class generate food randomly
to make snake eat food and increase size

What is the impact of the change?
food is generated randomly and snake is eating food and adding part to
it