Gave the ghost snake the ability to eat.
What changes you have made?
Gave the ghost snake the ability to eat food
Gave the ghost snake the ability to eat snake

Why this change is necessary?
for giving the ghost snake the ability to eat

What is the impact of the change?
The ghost snake able to eat snake and food.